requirements.txt
Django~=4.1
cryptography
djangorestframework
django-cors-headers
drf-jwt
pyjwt
requests
psycopg2
djangorestframework-simplejwt==5.0.0